+++ src/main.js
@@ -5,7 +5,7 @@ const filenames = [
   'js/babel',
   'js/coffeescript',
   'js/typescript',
-  'inline-html-ng'
+  'inline-html'
 ];
 
 module.exports = _.map(filenames, (x) => require('./' + x).default);
